requirements.txt
# pymilvus must be installed manually without milvus-lite
# pymilvus==2.4.0
grpcio==1.58.0
protobuf>=3.20.0
pandas>=1.2.4
environs<=9.5.0
openai==1.14.3
python-dotenv==1.0.1
python-docx==1.1.0
pymupdf==1.23.25
azure-storage-blob
minio>=7.0.0
pyarrow>=12.0.0
requests
scipy
ujson>=2.0.0
marshmallow>=3.13,<4.0
httpx==0.27.0
streamlit